Refactor hospital management to use NextJS dynamic routes
